Use `str_comp_filenames` to sort maps and demos again
This means that digit characters in filenames will be comparsed as numbers instead of being compared as individual digits.

The use of this function was previously (5f85c6b2014e3f86f489ade9eeac48c536c98fc1) removed as it did not sort names case insensitively, whereas now it does.