Revert "Delay disconnecting until after render call"
This reverts commit 59593e1a908bbb67c5a31c04a6ac0fdd2ac97395.

Delaying the disconnecting causes issues when the client automatically disconnects immediately before connecting to another server or starting demo playback.

It's not necessary to delay the disconnecting to deal with #6387, as #6589 is already enough. It's easier to revert this instead of rewriting the client so connecting and starting demo playback are also delayed.

Closes #6595.